split process class into growth, transfer and mortality classes
